requirements.txt
alpaca-py>=0.15.0
anthropic>=0.7.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
aiohttp>=3.8.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.1.0

# Additional utilities
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0